chore: split handler methods
