+++ BotDeScans.UnitTests/Specs/Features/Publish/Interaction/HandlerTests.cs
@@ -549,120 +549,4 @@ public class HandlerTests : UnitTest
             A.CallTo(() => publishStep.ExecuteAsync(A<State>.Ignored, cancellationToken)).MustHaveHappenedOnceExactly();
         }
     }
-
-    public class MergeStatesTests : HandlerTests
-    {
-        [Fact]
-        public void GivenParallelResultsShouldMergeNonNullLinkProperties()
-        {
-            var baseState = new State { MegaZipLink = "https://mega.nz/zip" };
-            var updatedState = new State
-            {
-                BoxZipLink = "https://box.com/zip",
-                Steps = new EnabledSteps([]),
-            };
-
-            var merged = Handler.MergeStates(baseState, updatedState);
-
-            merged.MegaZipLink.Should().Be("https://mega.nz/zip");
-            merged.BoxZipLink.Should().Be("https://box.com/zip");
-        }
-
-        [Fact]
-        public void GivenParallelConversionResultsShouldMergeFilePaths()
-        {
-            var baseState = new State { ZipFilePath = "/tmp/chapter.zip" };
-            var updatedState = new State
-            {
-                PdfFilePath = "/tmp/chapter.pdf",
-                Steps = new EnabledSteps([]),
-            };
-
-            var merged = Handler.MergeStates(baseState, updatedState);
-
-            merged.ZipFilePath.Should().Be("/tmp/chapter.zip");
-            merged.PdfFilePath.Should().Be("/tmp/chapter.pdf");
-        }
-
-        [Fact]
-        public void GivenUpdatedLinkShouldOverrideBaseLink()
-        {
-            var baseState = new State { MegaZipLink = "https://old.link" };
-            var updatedState = new State
-            {
-                MegaZipLink = "https://new.link",
-                Steps = new EnabledSteps([]),
-            };
-
-            var merged = Handler.MergeStates(baseState, updatedState);
-
-            merged.MegaZipLink.Should().Be("https://new.link");
-        }
-
-        [Fact]
-        public void GivenNullUpdatedLinkShouldPreserveBaseLink()
-        {
-            var baseState = new State { DriveZipLink = "https://drive.google.com/zip" };
-            var updatedState = new State
-            {
-                DriveZipLink = null,
-                Steps = new EnabledSteps([]),
-            };
-
-            var merged = Handler.MergeStates(baseState, updatedState);
-
-            merged.DriveZipLink.Should().Be("https://drive.google.com/zip");
-        }
-
-        [Fact]
-        public void GivenParallelSnapshotsShouldPreserveStepInfoFromBothSnapshots()
-        {
-            var stepA = A.Fake<IConversionStep>();
-            var stepB = A.Fake<IConversionStep>();
-
-            var baseSteps = new EnabledSteps(new Dictionary<IStep, StepInfo>
-            {
-                { stepA, new StepInfo(stepA) { Status = StepStatus.Success } },
-                { stepB, new StepInfo(stepB) { Status = StepStatus.QueuedForExecution } },
-            });
-
-            var updatedSteps = new EnabledSteps(new Dictionary<IStep, StepInfo>
-            {
-                { stepA, new StepInfo(stepA) { Status = StepStatus.QueuedForExecution } },
-                { stepB, new StepInfo(stepB) { Status = StepStatus.Success } },
-            });
-
-            var baseState = new State { Steps = baseSteps };
-            var updatedState = new State { Steps = updatedSteps };
-
-            var merged = Handler.MergeStates(baseState, updatedState);
-
-            using var _ = new AssertionScope();
-            merged.Steps[stepA].Status.Should().Be(StepStatus.Success);
-            merged.Steps[stepB].Status.Should().Be(StepStatus.Success);
-        }
-
-        [Fact]
-        public void GivenUpdatedStepInfoWithDifferentStatusShouldOverrideBaseStepInfo()
-        {
-            var step = A.Fake<IConversionStep>();
-
-            var baseSteps = new EnabledSteps(new Dictionary<IStep, StepInfo>
-            {
-                { step, new StepInfo(step) { Status = StepStatus.QueuedForExecution } },
-            });
-
-            var updatedSteps = new EnabledSteps(new Dictionary<IStep, StepInfo>
-            {
-                { step, new StepInfo(step) { Status = StepStatus.Success } },
-            });
-
-            var baseState = new State { Steps = baseSteps };
-            var updatedState = new State { Steps = updatedSteps };
-
-            var merged = Handler.MergeStates(baseState, updatedState);
-
-            merged.Steps[step].Status.Should().Be(StepStatus.Success);
-        }
-    }
 }
