+++ BotDeScans.UnitTests/Specs/Features/Publish/Interaction/StateTests.cs
@@ -1,4 +1,9 @@
 using BotDeScans.App.Features.Publish.Interaction;
+using BotDeScans.App.Features.Publish.Interaction.Models;
+using BotDeScans.App.Features.Publish.Interaction.Steps;
+using BotDeScans.App.Features.Publish.Interaction.Steps.Enums;
+using BotDeScans.App.Models.Entities.Enums;
+using FluentAssertions.Execution;
 
 namespace BotDeScans.UnitTests.Specs.Features.Publish.Interaction;
 
@@ -157,6 +162,110 @@ public class StateTests : UnitTest
             var newState = state with { BloggerLink = "new-link" };
             newState.BloggerLink.Should().Be("new-link");
             state.BloggerLink.Should().Be("blogger");
+        }
+    }
+
+    public class MergeWithMethod : StateTests
+    {
+        [Fact]
+        public void GivenParallelResultsShouldMergeNonNullLinkProperties()
+        {
+            var baseState = new State { MegaZipLink = "https://mega.nz/zip" };
+            var other = new State { BoxZipLink = "https://box.com/zip", Steps = new EnabledSteps([]) };
+
+            var merged = baseState.MergeWith(other);
+
+            merged.MegaZipLink.Should().Be("https://mega.nz/zip");
+            merged.BoxZipLink.Should().Be("https://box.com/zip");
+        }
+
+        [Fact]
+        public void GivenParallelConversionResultsShouldMergeFilePaths()
+        {
+            var baseState = new State { ZipFilePath = "/tmp/chapter.zip" };
+            var other = new State { PdfFilePath = "/tmp/chapter.pdf", Steps = new EnabledSteps([]) };
+
+            var merged = baseState.MergeWith(other);
+
+            merged.ZipFilePath.Should().Be("/tmp/chapter.zip");
+            merged.PdfFilePath.Should().Be("/tmp/chapter.pdf");
+        }
+
+        [Fact]
+        public void GivenUpdatedLinkShouldOverrideBaseLink()
+        {
+            var baseState = new State { MegaZipLink = "https://old.link" };
+            var other = new State { MegaZipLink = "https://new.link", Steps = new EnabledSteps([]) };
+
+            var merged = baseState.MergeWith(other);
+
+            merged.MegaZipLink.Should().Be("https://new.link");
+        }
+
+        [Fact]
+        public void GivenNullUpdatedLinkShouldPreserveBaseLink()
+        {
+            var baseState = new State { DriveZipLink = "https://drive.google.com/zip" };
+            var other = new State { DriveZipLink = null, Steps = new EnabledSteps([]) };
+
+            var merged = baseState.MergeWith(other);
+
+            merged.DriveZipLink.Should().Be("https://drive.google.com/zip");
+        }
+
+        [Fact]
+        public void GivenParallelSnapshotsShouldPreserveStepInfoFromBothSnapshots()
+        {
+            var stepA = A.Fake<IConversionStep>();
+            var stepB = A.Fake<IConversionStep>();
+
+            var baseState = new State
+            {
+                Steps = new EnabledSteps(new Dictionary<IStep, StepInfo>
+                {
+                    { stepA, new StepInfo(stepA) { Status = StepStatus.Success } },
+                    { stepB, new StepInfo(stepB) { Status = StepStatus.QueuedForExecution } },
+                })
+            };
+            var other = new State
+            {
+                Steps = new EnabledSteps(new Dictionary<IStep, StepInfo>
+                {
+                    { stepA, new StepInfo(stepA) { Status = StepStatus.QueuedForExecution } },
+                    { stepB, new StepInfo(stepB) { Status = StepStatus.Success } },
+                })
+            };
+
+            var merged = baseState.MergeWith(other);
+
+            using var _ = new AssertionScope();
+            merged.Steps[stepA].Status.Should().Be(StepStatus.Success);
+            merged.Steps[stepB].Status.Should().Be(StepStatus.Success);
+        }
+
+        [Fact]
+        public void GivenUpdatedStepInfoWithDifferentStatusShouldOverrideBaseStepInfo()
+        {
+            var step = A.Fake<IConversionStep>();
+
+            var baseState = new State
+            {
+                Steps = new EnabledSteps(new Dictionary<IStep, StepInfo>
+                {
+                    { step, new StepInfo(step) { Status = StepStatus.QueuedForExecution } },
+                })
+            };
+            var other = new State
+            {
+                Steps = new EnabledSteps(new Dictionary<IStep, StepInfo>
+                {
+                    { step, new StepInfo(step) { Status = StepStatus.Success } },
+                })
+            };
+
+            var merged = baseState.MergeWith(other);
+
+            merged.Steps[step].Status.Should().Be(StepStatus.Success);
         }
     }
 }
